requirements.txt
pandas
numpy
MetaTrader5
python-dotenv
requests
pytest
pytest-mock